+++ src/cpp/rest/dto/InternalStateDto.hpp
@@ -19,7 +19,7 @@ class InternalStateDto : public oatpp::DTO {
   DTO_FIELD_INFO(inference_duration_stats) {
     info->description = "Percentiles of time (in us) needed to infer one image";
   }
-  DTO_FIELD(List<Fields<Float64>>, inference_duration_stats,
+  DTO_FIELD(List<Fields<UInt32>>, inference_duration_stats,
             "inferenceDurationStats");
 };
